Move an_engine into astrometry2
