+++ skylib/astrometry2/__init__.py
@@ -15,6 +15,11 @@ from .main import (
     solve_field_glob,
 )
 
+try:  # pragma: no cover - optional dependency
+    from . import an_engine
+except Exception:  # pragma: no cover - missing optional dependency
+    an_engine = None
+
 __all__ = [
     "AstapBackend",
     "AstapConfig",
@@ -28,4 +33,5 @@ __all__ = [
     "SolveSolution",
     "solve_field",
     "solve_field_glob",
+    "an_engine",
 ]
